test(runtime): isolate retry resume control-flow regressions
